requirements.txt
ijson
simplejson
pandas>=1.5.0
numpy
scipy
matplotlib
scikit-learn
statsmodels
textblob
google-cloud-storage
google-cloud-secret-manager
tomli
bottleneck
gensim
more-itertools
altair
altair_saver
openpyxl
google-api-core 
google-auth 
google-auth-oauthlib 
google-cloud-bigquery
google-cloud-core 
google-resumable-media
googleapis-common-protos